requirements.txt
django==4
uvicorn
django-configurations
environs